Move areAllContactsVerified to backend file
