+++ src/renderer/components/helpers/ChatMethods.ts
@@ -260,20 +260,3 @@ export async function createDraftMessage(
 
   window.__reloadDraft && window.__reloadDraft()
 }
-
-/**
- * Returns true if all contacts of a given list are verified, otherwise false.
- */
-export async function areAllContactsVerified(
-  accountId: number,
-  contactIds: number[]
-): Promise<boolean> {
-  const contacts = await BackendRemote.rpc.getContactsByIds(
-    accountId,
-    contactIds
-  )
-
-  return !contactIds.some(contactId => {
-    return !contacts[contactId].isVerified
-  })
-}
